Renamed Manager.cs to MainWindow.cs and changed namespaces.
