+++ Forms/AddEvent.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Windows.Forms;
 
-namespace Timotheus
+namespace Timotheus.Forms
 {
     public partial class AddEvent : Form
     {
